backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Database
sqlalchemy==2.0.36
alembic==1.14.0
# psycopg2-binary==2.9.10  # Removed for SQLite compatibility
redis==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pyjwt==2.10.1

# AI & Computer Vision (Optional - install separately if needed)
# opencv-python==4.10.0.84
# opencv-contrib-python==4.10.0.84
# tensorflow==2.18.0
# face-recognition==1.3.0
# mtcnn==1.0.0
# deepface==0.0.93

# Image Processing
pillow==11.0.0
numpy==2.1.3
scikit-learn==1.6.0

# RFID Integration
pyserial==3.5
# pynfc==0.0.2  # Optional - may not be compatible

# Camera & Streaming (Optional)
# aiortc==1.10.0
# av==13.1.0
websockets==14.1

# Email & Notifications
sendgrid==6.11.0
python-dotenv==1.0.1
celery==5.4.0
kombu==5.4.2

# Validation & Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
python-dateutil==2.9.0
pytz==2024.2

# Monitoring & Logging
loguru==0.7.3
prometheus-client==0.21.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
